Fix autoloading of crux and zop-to-char commands.
Using :commands triggers use-package to defer loading, and the :config happens after loading.
Meaning that, with the previous config, those packages would have to been manually loaded before the
keybindings would be set.

This could be fixed by using :bind which sets up the keybindings eagerly but with autoloads for the
commands (what I've done here) or not using :command and eagerly loading the package plus using :config.