Fix before-save-hooks.
These are suppoed to be turned on only for certain buffers, as such I need to pass `t` as the fourth
argument to `add-hook`. Without that `local` set to true, once the add-hook executes, it is added to
the GLOBAL before-save-hook.

To wit: with that setting, I kept getting errors because lsp-format-buffer was getting called
on-save in my org files and other circumstances where lsp-mode was not on.

As it stands, I think if I turned lsp-mde OFF for a given buffer, the before-save hook would stay
on, but eh, that seems harder to fix.